# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "clipos API",
    "description": "Invite-only clip archive.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/clips/{id}/share": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Turn on the clip's public link (or get the existing one).",
        "operationId": "share_clip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Clip id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clip with its `shareUrl`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipView"
                }
              }
            }
          },
          "400": {
            "description": "Not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not yours",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "clips"
        ],
        "summary": "Revoke the clip's public link; anything already pasted stops working. Also for a clip\nin the trash.",
        "operationId": "unshare_clip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Clip id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clip without a link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipView"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not yours",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClipStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "processing",
          "ready",
          "failed"
        ]
      },
      "ClipView": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "gameId",
          "myPov",
          "status",
          "uploader",
          "isMine",
          "canEdit",
          "originalFilename",
          "originalBytes",
          "tags",
          "autoTags",
          "players",
          "reactions",
          "reactionCount",
          "createdAt",
          "teaser"
        ],
        "properties": {
          "autoTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags found by killfeed analysis, e.g. `4k`, `hs`, `smoke-kill`."
          },
          "canEdit": {
            "type": "boolean",
            "description": "The viewer may edit and delete it (uploader or admin)."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Set while it's in the trash (restorable for 7 days)."
          },
          "description": {
            "type": "string"
          },
          "durationMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why processing failed. Only shown to the uploader."
          },
          "failureReason": {
            "oneOf": [
              {
                "type": "string",
                "description": "Why a clip failed processing, for the uploader's failed screen. Worked out from the\nstored message (`clips.error`), whose permanent forms the worker writes with the\nconstants below.",
                "enum": [
                  "tooLong",
                  "notAVideo",
                  "unreadable",
                  "server"
                ]
              },
              {
                "type": "null"
              }
            ],
            "description": "The kind of failure, for the uploader's failed screen."
          },
          "fps": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "gameId": {
            "type": "string"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "heldUntil": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Saved for the show until then; only shown to the uploader."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isMine": {
            "type": "boolean"
          },
          "map": {
            "type": [
              "string",
              "null"
            ]
          },
          "multiKill": {
            "type": [
              "string",
              "null"
            ],
            "description": "The uploader's best round from killfeed analysis: `2k`, `3k`, `4k` or `ace`."
          },
          "myPov": {
            "type": "boolean"
          },
          "originalBytes": {
            "type": "integer",
            "format": "int64"
          },
          "originalFilename": {
            "type": "string"
          },
          "playbackUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Read SAS (2 h) once ready; only on the single-clip endpoints."
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "description": "Friends tagged as playing in it."
          },
          "posterUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Read SAS (2 h) once ready."
          },
          "reactionCount": {
            "type": "integer",
            "format": "int32"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionCount"
            }
          },
          "shareUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Public link while sharing is on; only shown to those who can manage it."
          },
          "status": {
            "$ref": "#/components/schemas/ClipStatus"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The uploader's tags (what editing the clip changes)."
          },
          "teaser": {
            "type": "boolean",
            "description": "Someone else's clip saved for the show: only its title, uploader, map and length,\nwith a blurred poster. No tags, killfeed or reactions, and no playback except for\npeople in the live show whose lineup has it (`GET /api/clips/{id}`, to preload it)."
          },
          "title": {
            "type": "string"
          },
          "uploader": {
            "$ref": "#/components/schemas/Uploader"
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Error payload for every non-2xx `/api` response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code: `bad_request`, `unauthorized`, `not_invited`,\n`account_disabled`, `forbidden`, `not_found`, `conflict`, `unsupported_media_type`,\n`too_large`, `bad_gateway` or `internal`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Never contains internals for 5xx responses."
          },
          "tied": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Ties",
                "description": "Only on the 409 from ending a show with a tied vote: the tied clips of every tied\ncategory, so the host breaks them all at once."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "Member": {
        "type": "object",
        "description": "A user as other members see them (no email).",
        "required": [
          "id",
          "handle",
          "displayName"
        ],
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "steamName": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ReactionCount": {
        "type": "object",
        "required": [
          "emoji",
          "count",
          "mine"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "emoji": {
            "type": "string"
          },
          "mine": {
            "type": "boolean",
            "description": "The viewer reacted with this emoji."
          }
        }
      },
...
      "Ties": {
        "type": "object",
        "description": "The tied clips of each category when `end` needs the host's pick; empty for a\ncategory that isn't tied.",
        "required": [
          "clip",
          "fail"
        ],
        "properties": {
          "clip": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "fail": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "Uploader": {
        "type": "object",
        "required": [
          "handle",
          "displayName"
        ],
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}